requirements.txt
Django==3.0.14
psycopg2-binary
mysqlclient==2.1.0
requests==2.32.3
celery
redis==4.6.0
django-redis==3.5.2
django-celery-results==1.0.1
django-celery-beat
django-recaptcha
django-cache-url==3.0.0
django-mathfilters==1.0.0
markdown==3.0.1
django-filter
uwsgi
dj-database-url==0.5.0
django-silk
django-cors-middleware==1.3.1
djangorestframework
djangorestframework-api-key==2.0.0
drf-yasg==1.20.0
kafka-python
sentry-sdk
pymongo
cryptography==43.0.0
jwcrypto>=0.8
APScheduler>=3.6.4
django-constance[database]
django-debug-toolbar
django-debug-toolbar-request-history
django-cors-headers==3.5.0
openpyxl==3.0.9
django-currentuser==0.5.3
gunicorn==20.1.0
mongoengine==0.24.2
PyJWT==2.9.0
django-rest-swagger==2.2.0
djangorestframework-simplejwt==5.2.2
